add basket component and routing; refactor form handling and UI updates

+++ src/pages/Home/Home.jsx
@@ -15,7 +15,7 @@ export const Home = () => {
     setName(carNames);
   }, []);
 
-  function HandleForm(e) {
+  const HandleForm = (e) => {
     e.preventDefault();
     if (!selectedCar || !duedate || !returndate) {
       alert("Zəhmət olmasa bütün xanaları doldurun!");
@@ -23,7 +23,7 @@ export const Home = () => {
     }
     const start = new Date(duedate);
     const end = new Date(returndate);
-    const diffDays = Math.ceil((end - start) / (1000 * 60 * 60 * 24));
+    const diffDays = (end - start) / (1000 * 60 * 60 * 24);
     if (diffDays <= 0) {
       alert("Qaytarılacaq tarix götürüləcək tarixdən sonra olmalıdır!");
       return;
@@ -32,17 +32,26 @@ export const Home = () => {
     if (car) {
       setPrice(car.price * diffDays);
     }
-  }
+    setSelectedCar("");
+    setDueDate("");
+    setReturnDate("");
+  };
 
-  function HandleClick(e) {
-  const car = Cars.find((c) => c.id === e);
-  const existing = JSON.parse(localStorage.getItem("selectedCars")) || [];
+  const handleClick = (e) => {
+    const car = Cars.find((c) => c.id === e);
 
-  const updated = [...existing, car];
+    const existing = JSON.parse(localStorage.getItem("selectedCars")) || [];
 
-  localStorage.setItem("selectedCars", JSON.stringify(updated));
-  }
+    const exists = existing.find((item) => item.id === e);
 
+    if (exists) {
+      return;
+    }
+
+    const updated = [...existing, car];
+
+    localStorage.setItem("selectedCars", JSON.stringify(updated));
+  };
 
   return (
     <div>
@@ -90,6 +99,7 @@ export const Home = () => {
             <input
               type="date"
               name=""
+              value={duedate}
               onChange={(e) => setDueDate(e.target.value)}
               className="outline-none border text-gray-400 border-gray-300 rounded-sm w-50 px-2 py-1 my-2"
             />
@@ -106,17 +116,15 @@ export const Home = () => {
               className="outline-none border text-gray-400 border-gray-300 rounded-sm w-50 px-2 py-1 my-2 "
             />
           </div>
-        
+
           <button
             type="submit"
             className=" text-white bg-blue-600 px-4 h-12 rounded-lg"
           >
             Qiymətə bax
           </button>
-            <div className="flex flex-col">
-            <label className="text-gray-800 font-medium">
-              Qiymət:
-            </label>
+          <div className="flex flex-col">
+            <label className="text-gray-800 font-medium">Qiymət:</label>
             <span className="min-w-20 text-center border text-gray-400 border-gray-300 rounded-sm px-2 py-1 my-2 ">
               {price > 0 ? `${price} AZN` : "-"}
             </span>
@@ -132,44 +140,56 @@ export const Home = () => {
         </div>
 
         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-6 mt-6">
-          {Cars?.map((car) => (
-            <div
-              key={car.id}
-              className="bg-white rounded-2xl border border-gray-200 p-4 w-[300px] shadow-sm hover:shadow-md transition"
-            >
-              <div className="flex justify-center mb-4">
-                <img
-                  src={car.image}
-                  alt="Audi A6"
-                  className="h-32 object-contain"
-                />
-              </div>
-              <h2 className="text-lg font-semibold text-gray-800 mb-2">
-                {car.name}
-              </h2>
-              <div className="flex items-center gap-4 text-sm text-gray-500 mb-4">
-                <span className="flex items-center gap-2 text-gray-500">
-                  <FaCogs className="text-gray-600" /> {car.transmission}
-                </span>
-                <span className="flex items-center gap-2 text-gray-500">
-                  <FaGasPump className="text-gray-600" /> {car.fuel}
-                </span>
-                <span className="flex items-center gap-2 text-gray-500">
-                  <FaUserFriends className="text-gray-600" /> {car.seats}
-                </span>
-              </div>
-              <div className="flex items-center justify-between">
-                <p className="text-lg font-semibold text-gray-800">
-                  {car.price} AZN
-                  <span className="text-gray-400 text-sm">/ gün</span>
-                </p>
+          {Cars?.map((car) => { 
+            const existing =
+              JSON.parse(localStorage.getItem("selectedCars")) || [];
+            const exists = existing.find((item) => item.id === car.id);
+            return (
+              <div
+                key={car.id}
+                className="bg-white rounded-2xl border border-gray-200 p-4 w-[300px] shadow-sm hover:shadow-md transition"
+              >
+                <div className="flex justify-center mb-4">
+                  <img
+                    src={car.image}
+                    alt="Audi A6"
+                    className="h-32 object-contain"
+                  />
+                </div>
+                <h2 className="text-lg font-semibold text-gray-800 mb-2">
+                  {car.name}
+                </h2>
+                <div className="flex items-center gap-4 text-sm text-gray-500 mb-4">
+                  <span className="flex items-center gap-2 text-gray-500">
+                    <FaCogs className="text-gray-600" /> {car.transmission}
+                  </span>
+                  <span className="flex items-center gap-2 text-gray-500">
+                    <FaGasPump className="text-gray-600" /> {car.fuel}
+                  </span>
+                  <span className="flex items-center gap-2 text-gray-500">
+                    <FaUserFriends className="text-gray-600" /> {car.seats}
+                  </span>
+                </div>
+                <div className="flex items-center justify-between">
+                  <p className="text-lg font-semibold text-gray-800">
+                    {car.price} AZN
+                    <span className="text-gray-400 text-sm">/ gün</span>
+                  </p>
 
-                <button className="border border-blue-500 text-blue-500 px-8 py-1.5 rounded-lg hover:bg-blue-500 hover:text-white transition" onClick={()=>HandleClick(car.id)}>
-                  Seç
-                </button>
+                  <button
+                    className={`border border-blue-500 text-blue-500 px-8 py-1.5 rounded-lg hover:bg-blue-500 hover:text-white transition  ${
+                      exists
+                        ? "bg-blue-500 text-white"
+                        : "bg-white text-blue-500 hover:bg-blue-500 hover:text-white"
+                    }`}
+                    onClick={() => handleClick(car.id)}
+                  >
+                    Səbətə at
+                  </button>
+                </div>
               </div>
-            </div>
-          ))}
+            );
+          })}
         </div>
       </section>
     </div>
